Do not use event handler for loading composable nodes
Fixes #114.

Due to the asynchronous nature of the LoadComposableNodes action, an event handler causes the launch configuration is popped if ComposableNodeContainer appears inside a group or include action.
It seems to me we can simply return the load node action, which will get executed after the ComposableNodeContainer action. The use of an event handler is vestigial of a refactoring done in https://github.com/ros2/launch_ros/pull/16, and doesn't appear to be necessary.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>