Support Python 3.8 importlib.metadata, declare dependency
The importlib_metadata package is a backport of the importlib.metadata
module from Python 3.8. We should prefer the one coming from Python
itself, and fall back to the backport if Python doesn't provide it.

Since this could mean an additional dependency on some platforms, we
should also declare that dependency.

Signed-off-by: Scott K Logan <logans@cottsay.net>