Inherit markers from generate_test_description
As it stands, pytest sees rostest tests as a single testing node with no
markings. It would be useful if normal pytest markings could be used on
the generate_test_description function (where @pytest.mark.rostest and
@launch_testing.parametrize are currently specified).

Notably, this will always mark the test nodes with 'rostest', and will
allow the use of very useful marking like 'xfail', 'skip', and 'skipif'.

Signed-off-by: Scott K Logan <logans@cottsay.net>