Give the option to inject a quality of service profile (#493)
Signed-off-by: Giorgio Pintaudi <pintaudi@axelspace.com>

When testing subscribers and  publishers, usually we do not want to miss any message due to race issues or other network instabilities. So we give the option to the user to specify a QoS profile best suited for the test scenario. The default is to set the durability to transient local, so that the publisher do not drop any packages in case of late joining subscribers.